Refactor to allow usage of internal api in tests
